- aligned qr dependency w/ files app - aligned qr image with files - Exiting qr directs back to ServerSelectionActivity
