+++ gradle/verification-metadata.xml
@@ -301,6 +301,7 @@
          <trusted-key id="CBD134820A0E04D230AEC5D5D9B2BC72A3754BEA" group="^com[.]github[.]lukaspili($|([.].*))" regex="true"/>
          <trusted-key id="CD5464315F0B98C77E6E8ECD9DAADC1C9FCC82D0" group="org.apache.commons" name="commons-lang3"/>
          <trusted-key id="CE0E1BC443809579D8718F4FB943F5CB616566CD" group="de.mannodermaus.retrofit2" name="converter-logansquare" version="1.4.1"/>
+         <trusted-key id="CE3285F320685193D11FEA01F6CE9695C9318406" group="com.google.zxing"/>
          <trusted-key id="CE8075A251547BEE249BC151A2115AE15F6B8B72" group="org.apache.commons"/>
          <trusted-key id="CFC10B69382CBCF5387E51484ECE492B63E38ACF" group="com.h3xstream.findsecbugs"/>
          <trusted-key id="D410F787672BA63BC2DD192B0C907617691418AE" group="org.greenrobot"/>
@@ -11961,6 +11962,14 @@
          </artifact>
          <artifact name="android-talk-webrtc-132.6834.0.pom">
             <sha256 value="70693e430f2a7e166690af2eeabf5c94109e68818cc66d453b1a43ee60978ea0" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+      </component>
+      <component group="com.github.nextcloud-deps" name="qrcodescanner" version="0.1.2.4">
+         <artifact name="qrcodescanner-0.1.2.4.aar">
+            <sha256 value="b286128792cc04f59b0defa2c937c86d9e2fc824a8011b9af9eea7fd0ea84303" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+         <artifact name="qrcodescanner-0.1.2.4.module">
+            <sha256 value="fbdd13cbf3f0288a063602d0eff83572039f8e623612576da3e10a49349c60a9" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
       </component>
       <component group="com.github.nextcloud-deps.hwsecurity" name="hwsecurity" version="4.1.0-patch2">
